Move authorize button controller to admin subfolder
